+++ peppy/project.py
@@ -746,8 +746,9 @@ class Project(PathExAttMap):
             _LOGGER.warning(no_metadata_msg.format(CFG_SAMPLE_TABLE_KEY))
             self[SAMPLE_DF_KEY] = None
         if CFG_SUBSAMPLE_TABLE_KEY in self[CONFIG_KEY]:
-            sst = make_list(self[CONFIG_KEY][CFG_SUBSAMPLE_TABLE_KEY], str)
-            self[SUBSAMPLE_DF_KEY] = [_read_tab(x) for x in sst]
+            if self[CONFIG_KEY][CFG_SUBSAMPLE_TABLE_KEY] is not None:
+                sst = make_list(self[CONFIG_KEY][CFG_SUBSAMPLE_TABLE_KEY], str)
+                self[SUBSAMPLE_DF_KEY] = [_read_tab(x) for x in sst]
         else:
             _LOGGER.debug(no_metadata_msg.format(CFG_SUBSAMPLE_TABLE_KEY))
             self[SUBSAMPLE_DF_KEY] = None
@@ -814,9 +815,12 @@ class Project(PathExAttMap):
                 #  reformatting or drop the old cfg format altogether
                 if k_from in ["implied_attributes", "implied_columns"]:
                     raise NotImplementedError(
-                        "Implications reformatting is not yet implemented. "
-                        "Edit the config file manually to comply with "
-                        "PEP 2.0.0 specification: http://pep.databio.org")
+                        "The attribute implications section ({}) follows the "
+                        "old format. Reformatting is not implemented. Edit the "
+                        "config file manually (add '{}.{}') to comply with PEP "
+                        "2.0.0 specification: http://pep.databio.org/en/latest/"
+                        "specification/#sample-modifier-imply".format(
+                            k_from, SAMPLE_MODS_KEY, IMPLIED_KEY))
                 map.setdefault(SAMPLE_MODS_KEY, PathExAttMap())
                 if isinstance(k_to, list):
                     if k_to[0] in map[SAMPLE_MODS_KEY]: