+++ peppy/utils.py
@@ -36,9 +36,9 @@ def grab_project_data(prj):
         return {}
 
     try:
-        data = prj[CONFIG_KEY]
+        return prj[CONFIG_KEY].to_dict()
     except KeyError:
-        _LOGGER.debug("Project lacks section {}, skipping".format(CONFIG_KEY))
+        raise KeyError("Project lacks section '{}'".format(CONFIG_KEY))
     return data